[이지은] feat: 게시글 등록 페이지 ui 수정 [BRP]

+++ jieunLee/backend/src/main/java/com/example/backend/service/BoardServiceImpl.java
@@ -74,10 +74,14 @@ public class BoardServiceImpl implements BoardService {
         Board board = repository.findById(boardNo).orElseThrow();
 
         Optional<Comment> maybeComment = commentRepository.findByBoard(board);
-        commentRepository.deleteById(maybeComment.get().getCommentNo());
+        if(!maybeComment.isEmpty()) {
+            commentRepository.deleteById(maybeComment.get().getCommentNo());
+        }
 
         Optional<Likes> maybeLikes = likesRepository.findByBoard(board);
-        likesRepository.deleteById(maybeLikes.get().getLikedNo());
+        if(!maybeLikes.isEmpty()) {
+            likesRepository.deleteById(maybeLikes.get().getLikedNo());
+        }
 
         repository.deleteById(Long.valueOf(boardNo));
     }
